+++ wms-praktyki-yosi-api/wms-praktyki-yosi-api/Controllers/ProductsContoller.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using System.Security.Claims;
 using wms_praktyki_yosi_api.Enitities;
 using wms_praktyki_yosi_api.Models;
 using wms_praktyki_yosi_api.Services;
@@ -13,8 +14,11 @@ namespace wms_praktyki_yosi_api.Controllers
     {
 
         public readonly IProductService _productService;
-        public ProductsController(IProductService productService) {
+        private readonly IAccountService _accountService;
+
+        public ProductsController(IProductService productService, IAccountService accountService) {
             _productService = productService;
+            _accountService = accountService;
         }
 
         [HttpGet]
@@ -25,8 +29,11 @@ namespace wms_praktyki_yosi_api.Controllers
         }
 
         [HttpGet("{id}")]
-        public ActionResult<Product> Get([FromRoute]int id)
+        public async Task<ActionResult<Product>> Get([FromRoute]int id)
         {
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
             var product = _productService.GetById(id);
             if (product == null)
             {
@@ -36,9 +43,12 @@ namespace wms_praktyki_yosi_api.Controllers
         }
         [HttpPut("{id}")]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult Update([FromRoute] int id, [FromBody]ProductDto dto)
+        public async Task<ActionResult> Update([FromRoute] int id, [FromBody]ProductDto dto)
         {
-            if(!ModelState.IsValid)
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
+            if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
@@ -51,8 +61,11 @@ namespace wms_praktyki_yosi_api.Controllers
 
         [HttpDelete("{id}")]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult DeleteProduct([FromRoute]int id)
+        public async Task<ActionResult> DeleteProduct([FromRoute]int id)
         {
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
             var delete = _productService.RemoveProduct(id);
             if (!delete)
             {
@@ -62,15 +75,33 @@ namespace wms_praktyki_yosi_api.Controllers
         }
         [HttpPost]
         [Authorize(Roles = "Admin,Moderator")]
-        public ActionResult AddProduct([FromBody] ProductDto dto)
+        public async Task<ActionResult> AddProduct([FromBody] ProductDto dto)
         {
-           
-            if(!ModelState.IsValid)
+            if (!await UserIsAuthorized(User))
+                return Unauthorized();
+
+            if (!ModelState.IsValid)
             {
                 return BadRequest(2);
             }
             var productId = _productService.AddNewProduct(dto);
             return Created("/api/products/" + productId, null);
+        }
+        private async Task<bool> UserIsAuthorized(ClaimsPrincipal user)
+        {
+            var userEmail = user.FindFirst(ClaimTypes.Name);
+            if (userEmail is null)
+                return false;
+
+            try
+            {
+                var userInfo = await _accountService.GetUserInfo(userEmail.Value);
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
         }
     }
